+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/dao/PersistenceValueDaoImpl.java
@@ -10,9 +10,14 @@ import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
 import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
 import com.dwarfeng.subgrade.sdk.bean.key.HibernateLongIdKey;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
+import org.apache.commons.lang3.tuple.Pair;
+import org.hibernate.criterion.DetachedCriteria;
+import org.hibernate.criterion.Order;
+import org.hibernate.criterion.Restrictions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +28,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
@@ -41,20 +47,22 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
     private HibernateBatchWriteDao<PersistenceValue, HibernatePersistenceValue> batchWriteDao;
     @Autowired
     private HibernateTemplate hibernateTemplate;
+    @Autowired
+    private BeanTransformer<PersistenceValue, HibernatePersistenceValue> beanTransformer;
     @SuppressWarnings("FieldMayBeFinal")
     @Autowired(required = false)
-    private List<PersistenceValueNSQLQuery> nsqlGenerators = Collections.emptyList();
+    private List<PersistenceValueNSQLQuery> nsqlQueries = Collections.emptyList();
 
     @Value("${hibernate.accelerate.using_native_sql}")
     private boolean usingNativeSQL;
     @Value("${hibernate.dialect}")
     private String hibernateDialect;
 
-    private PersistenceValueNSQLQuery nsqlGenerator = null;
+    private PersistenceValueNSQLQuery nsqlQuery = null;
 
     @PostConstruct
     public void init() {
-        nsqlGenerator = nsqlGenerators.stream()
+        nsqlQuery = nsqlQueries.stream()
                 .filter(generator -> generator.supportType(hibernateDialect)).findAny().orElse(null);
     }
 
@@ -162,7 +170,7 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
     public List<PersistenceValue> lookup(String preset, Object[] objs) throws DaoException {
         try {
             if (Objects.equals(PersistenceValueMaintainService.CHILD_FOR_POINT_BETWEEN, preset) && usingNativeSQL) {
-                if (Objects.isNull(nsqlGenerator)) {
+                if (Objects.isNull(nsqlQuery)) {
                     LOGGER.warn("指定的 hibernateDialect: " + hibernateDialect + ", 不受支持, 将不会使用原生SQL进行查询");
                     return presetLookupDao.lookup(preset, objs);
                 }
@@ -170,7 +178,7 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
                 List<PersistenceValue> persistenceValues = hibernateTemplate.executeWithNativeSession(
                         session -> session.doReturningWork(connection -> {
                             try {
-                                return nsqlGenerator.lookupPersistence(connection, objs);
+                                return nsqlQuery.lookupPersistence(connection, objs);
                             } catch (Exception e) {
                                 LOGGER.warn("原生SQL查询返回异常", e);
                                 return null;
@@ -199,7 +207,7 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
     public List<PersistenceValue> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
         try {
             if (Objects.equals(PersistenceValueMaintainService.CHILD_FOR_POINT_BETWEEN, preset) && usingNativeSQL) {
-                if (Objects.isNull(nsqlGenerator)) {
+                if (Objects.isNull(nsqlQuery)) {
                     LOGGER.warn("指定的 hibernateDialect: " + hibernateDialect + ", 不受支持, 将不会使用原生SQL进行查询");
                     return presetLookupDao.lookup(preset, objs, pagingInfo);
                 }
@@ -207,7 +215,7 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
                 List<PersistenceValue> persistenceValues = hibernateTemplate.executeWithNativeSession(
                         session -> session.doReturningWork(connection -> {
                             try {
-                                return nsqlGenerator.lookupPersistence(connection, objs, pagingInfo);
+                                return nsqlQuery.lookupPersistence(connection, objs, pagingInfo);
                             } catch (Exception e) {
                                 LOGGER.warn("原生SQL查询返回异常", e);
                                 return null;
@@ -236,7 +244,7 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
     public int lookupCount(String preset, Object[] objs) throws DaoException {
         try {
             if (Objects.equals(PersistenceValueMaintainService.CHILD_FOR_POINT_BETWEEN, preset) && usingNativeSQL) {
-                if (Objects.isNull(nsqlGenerator)) {
+                if (Objects.isNull(nsqlQuery)) {
                     LOGGER.warn("指定的 hibernateDialect: " + hibernateDialect + ", 不受支持, 将不会使用原生SQL进行查询");
                     return presetLookupDao.lookupCount(preset, objs);
                 }
@@ -244,7 +252,7 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
                 Integer count = hibernateTemplate.executeWithNativeSession(
                         session -> session.doReturningWork(connection -> {
                             try {
-                                return nsqlGenerator.lookupPersistenceCount(connection, objs);
+                                return nsqlQuery.lookupPersistenceCount(connection, objs);
                             } catch (Exception e) {
                                 LOGGER.warn("原生SQL查询返回异常", e);
                                 return null;
@@ -279,5 +287,53 @@ public class PersistenceValueDaoImpl implements PersistenceValueDao {
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchWrite(List<PersistenceValue> elements) throws DaoException {
         batchWriteDao.batchWrite(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public PersistenceValue previous(LongIdKey pointKey, Date date) throws DaoException {
+        try {
+            if (Objects.isNull(nsqlQuery)) {
+                LOGGER.warn("指定的 hibernateDialect: " + hibernateDialect + ", 不受支持, 将不会使用原生SQL进行查询");
+                return previousByCriteria(pointKey, date);
+            }
+            LOGGER.debug("使用原生SQL进行查询...");
+            Pair<PersistenceValue, Exception> queryInfo = hibernateTemplate.executeWithNativeSession(
+                    session -> session.doReturningWork(connection -> {
+                        PersistenceValue persistenceValue = null;
+                        Exception exception = null;
+                        try {
+                            persistenceValue = nsqlQuery.previous(connection, pointKey, date);
+                        } catch (Exception e) {
+                            LOGGER.warn("原生SQL查询返回异常", e);
+                            exception = e;
+                        }
+                        return Pair.of(persistenceValue, exception);
+                    })
+            );
+            assert queryInfo != null;
+            if (Objects.isNull(queryInfo.getRight())) {
+                return queryInfo.getLeft();
+            } else {
+                LOGGER.warn("原生SQL查询返回值无效, 不使用原生SQL再次进行查询...");
+                return previousByCriteria(pointKey, date);
+            }
+        } catch (Exception e) {
+            throw new DaoException(e);
+        }
+    }
+
+    public PersistenceValue previousByCriteria(LongIdKey pointKey, Date date) {
+        DetachedCriteria criteria = DetachedCriteria.forClass(HibernatePersistenceValue.class);
+        if (Objects.isNull(pointKey)) {
+            criteria.add(Restrictions.isNull("pointLongId"));
+        } else {
+            criteria.add(Restrictions.eq("pointLongId", pointKey.getLongId()));
+        }
+        criteria.add(Restrictions.lt("happenedDate", date));
+        criteria.addOrder(Order.desc("happenedDate"));
+        return hibernateTemplate.findByCriteria(criteria, 0, 1).stream().findFirst()
+                .map(value -> beanTransformer.reverseTransform((HibernatePersistenceValue) value)).orElse(null);
     }
 }