+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/dao/TriggeredValueDaoImpl.java
@@ -10,9 +10,14 @@ import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
 import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
 import com.dwarfeng.subgrade.sdk.bean.key.HibernateLongIdKey;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
+import org.apache.commons.lang3.tuple.Pair;
+import org.hibernate.criterion.DetachedCriteria;
+import org.hibernate.criterion.Order;
+import org.hibernate.criterion.Restrictions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +28,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.PostConstruct;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
@@ -41,9 +47,11 @@ public class TriggeredValueDaoImpl implements TriggeredValueDao {
     private HibernateBatchWriteDao<TriggeredValue, HibernateTriggeredValue> batchWriteDao;
     @Autowired
     private HibernateTemplate hibernateTemplate;
+    @Autowired
+    private BeanTransformer<TriggeredValue, HibernateTriggeredValue> beanTransformer;
     @SuppressWarnings("FieldMayBeFinal")
     @Autowired(required = false)
-    private List<TriggeredValueNSQLQuery> nsqlGenerators = Collections.emptyList();
+    private List<TriggeredValueNSQLQuery> nsqlQueries = Collections.emptyList();
 
     @Value("${hibernate.accelerate.using_native_sql}")
     private boolean usingNativeSQL;
@@ -54,7 +62,7 @@ public class TriggeredValueDaoImpl implements TriggeredValueDao {
 
     @PostConstruct
     public void init() {
-        nsqlQuery = nsqlGenerators.stream()
+        nsqlQuery = nsqlQueries.stream()
                 .filter(generator -> generator.supportType(hibernateDialect)).findAny().orElse(null);
     }
 
@@ -345,5 +353,53 @@ public class TriggeredValueDaoImpl implements TriggeredValueDao {
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
     public void batchWrite(List<TriggeredValue> elements) throws DaoException {
         batchWriteDao.batchWrite(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public TriggeredValue previous(LongIdKey pointKey, Date date) throws DaoException {
+        try {
+            if (Objects.isNull(nsqlQuery)) {
+                LOGGER.warn("指定的 hibernateDialect: " + hibernateDialect + ", 不受支持, 将不会使用原生SQL进行查询");
+                return previousByCriteria(pointKey, date);
+            }
+            LOGGER.debug("使用原生SQL进行查询...");
+            Pair<TriggeredValue, Exception> queryInfo = hibernateTemplate.executeWithNativeSession(
+                    session -> session.doReturningWork(connection -> {
+                        TriggeredValue triggeredValue = null;
+                        Exception exception = null;
+                        try {
+                            triggeredValue = nsqlQuery.previous(connection, pointKey, date);
+                        } catch (Exception e) {
+                            LOGGER.warn("原生SQL查询返回异常", e);
+                            exception = e;
+                        }
+                        return Pair.of(triggeredValue, exception);
+                    })
+            );
+            assert queryInfo != null;
+            if (Objects.isNull(queryInfo.getRight())) {
+                return queryInfo.getLeft();
+            } else {
+                LOGGER.warn("原生SQL查询返回值无效, 不使用原生SQL再次进行查询...");
+                return previousByCriteria(pointKey, date);
+            }
+        } catch (Exception e) {
+            throw new DaoException(e);
+        }
+    }
+
+    public TriggeredValue previousByCriteria(LongIdKey pointKey, Date date) {
+        DetachedCriteria criteria = DetachedCriteria.forClass(HibernateTriggeredValue.class);
+        if (Objects.isNull(pointKey)) {
+            criteria.add(Restrictions.isNull("pointLongId"));
+        } else {
+            criteria.add(Restrictions.eq("pointLongId", pointKey.getLongId()));
+        }
+        criteria.add(Restrictions.lt("happenedDate", date));
+        criteria.addOrder(Order.desc("happenedDate"));
+        return hibernateTemplate.findByCriteria(criteria, 0, 1).stream().findFirst()
+                .map(value -> beanTransformer.reverseTransform((HibernateTriggeredValue) value)).orElse(null);
     }
 }