+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/dao/TriggeredValueNSQLQuery.java
@@ -2,10 +2,12 @@ package com.dwarfeng.fdr.impl.dao;
 
 import com.dwarfeng.fdr.stack.bean.entity.TriggeredValue;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
 import org.springframework.lang.NonNull;
 
 import java.sql.Connection;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -27,4 +29,6 @@ public interface TriggeredValueNSQLQuery extends NSQLQuery {
     List<TriggeredValue> lookupTriggeredForTrigger(@NonNull Connection connection, Object[] objs, PagingInfo pagingInfo) throws DaoException;
 
     Integer lookupTriggeredCountForTrigger(@NonNull Connection connection, Object[] objs) throws DaoException;
+
+    TriggeredValue previous(@NonNull Connection connection, LongIdKey pointKey, Date date) throws DaoException;
 }