+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/dao/PersistenceValueDao.java
@@ -6,6 +6,9 @@ import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
 import com.dwarfeng.subgrade.stack.dao.BatchWriteDao;
 import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
 import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
+import com.dwarfeng.subgrade.stack.exception.DaoException;
+
+import java.util.Date;
 
 /**
  * 持久化数据数据访问层。
@@ -15,4 +18,17 @@ import com.dwarfeng.subgrade.stack.dao.PresetLookupDao;
  */
 public interface PersistenceValueDao extends BatchBaseDao<LongIdKey, PersistenceValue>,
         EntireLookupDao<PersistenceValue>, PresetLookupDao<PersistenceValue>, BatchWriteDao<PersistenceValue> {
+
+    /**
+     * 获取属于指定数据点下的距离指定日期之前最后的持久化数据。
+     * <p>
+     * 获取的数据可以是 <code>null</code>。
+     *
+     * @param pointKey 指定的数据点。
+     * @param date     指定的日期
+     * @return 属于指定数据点下的距离指定日期之前最后的持久化数据，可以是 null。
+     * @throws DaoException 数据访问层异常。
+     * @since 1.9.0
+     */
+    PersistenceValue previous(LongIdKey pointKey, Date date) throws DaoException;
 }